Added a test for history in clickhouse-client
